Add algorithm::cliques, GraphConstructible trait and Graph::add_edges_from.
Added an iterator (generator-like) for the maximal cliques of a graph
based on Bron-Kerbosch - Conte-Tomita.

Moved self-allocating functions out of the Graph trait to a
GraphConstructible trait as a first step to make Graph an object-safe trait.
This imposed changes a bit everywhere to add the relevant import.

Add Graph::add_edges_from to ease construction of test graphs.

Also reformat a bunch of the files, sorry.